requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anthropic==0.45.2
antlr4-python3-runtime==4.9.3
anyio==4.8.0
asgiref==3.8.1
backoff==2.2.1
beautifulsoup4==4.12.3
cachetools==5.5.1
certifi==2024.12.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
coloredlogs==15.0.1
contourpy==1.3.1
cryptography==44.0.0
cycler==0.12.1
dataclasses-json==0.6.7
defusedxml==0.7.1
Deprecated==1.2.18
distro==1.9.0
dj-rest-auth==7.0.1
Django==5.1.5
django-allauth==65.3.1
django-cors-headers==4.6.0
django-rest-framework==0.1.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
effdet==0.4.1
emoji==2.14.1
et_xmlfile==2.0.0
eval_type_backport==0.2.2
filelock==3.17.0
filetype==1.2.0
flatbuffers==25.1.24
fonttools==4.55.8
fsspec==2024.12.0
google-api-core==2.24.1
google-auth==2.38.0
google-cloud-vision==3.9.0
googleapis-common-protos==1.66.0
grpcio==1.70.0
grpcio-status==1.70.0
gunicorn==23.0.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.28.1
humanfriendly==10.0
idna==3.10
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
kiwisolver==1.4.8
langchain-anthropic==0.3.5
langchain-core==0.3.33
langdetect==1.0.9
langgraph==0.2.68
langgraph-checkpoint==2.0.10
langgraph-sdk==0.1.51
langsmith==0.3.3
lxml==5.3.0
Markdown==3.7
MarkupSafe==3.0.2
marshmallow==3.26.0
matplotlib==3.10.0
mpmath==1.3.0
msgpack==1.1.0
mypy-extensions==1.0.0
neo4j==5.27.0
neomodel==5.4.2
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy==1.26.4
oauthlib==3.2.2
olefile==0.47
omegaconf==2.3.0
onnx==1.17.0
onnxruntime==1.20.1
opencv-python==4.11.0.86
openpyxl==3.1.5
orjson==3.10.15
packaging==24.2
pandas==2.2.3
pdf2image==1.17.0
pdfminer.six==20240706
pi_heif==0.21.0
pikepdf==9.5.1
pillow==11.1.0
proto-plus==1.26.0
protobuf==5.29.3
psutil==6.1.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycocotools==2.0.8
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
pypandoc==1.15
pyparsing==3.2.1
pypdf==5.2.0
pypdfium2==4.30.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-docx==1.1.2
python-iso639==2025.1.28
python-magic==0.4.27
python-multipart==0.0.20
python-oxmsg==0.0.1
python-pptx==1.0.2
pytz==2024.2
PyYAML==6.0.2
RapidFuzz==3.11.0
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rsa==4.9
safetensors==0.5.2
scipy==1.15.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.3
sympy==1.13.3
tenacity==9.0.0
timm==1.0.14
tokenizers==0.21.0
torch==2.2.2
torchvision==0.17.2
tqdm==4.67.1
transformers==4.48.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
unstructured==0.16.17
unstructured-client==0.29.0
unstructured-inference==0.8.6
unstructured.pytesseract==0.3.13
urllib3==2.3.0
webencodings==0.5.1
whitenoise==6.8.2
wrapt==1.17.2
xlrd==2.0.1
XlsxWriter==3.2.2
zstandard==0.23.0
